Move FAQ back to components
