+++ components/FAQ/style.module.scss
@@ -17,6 +17,7 @@
   width: 100%;
   display: flex;
   justify-content: space-between;
+  align-items: center;
 
   &:focus,
   &:hover {
